+++ src/r_typed_data.rs
@@ -67,7 +67,7 @@ impl RTypedData {
     /// # Examples
     ///
     /// ```
-    /// use magnus::{define_class, prelude::*, RTypedData};
+    /// use magnus::{class, define_class, prelude::*, RTypedData};
     /// # let _cleanup = unsafe { magnus::embed::init() };
     ///
     /// #[magnus::wrap(class = "Point")]
@@ -76,7 +76,7 @@ impl RTypedData {
     ///     y: isize,
     /// }
     ///
-    /// let point_class = define_class("Point", Default::default()).unwrap();
+    /// let point_class = define_class("Point", class::object()).unwrap();
     ///
     /// let value = RTypedData::wrap(Point { x: 4, y: 2 });
     /// assert!(value.is_kind_of(point_class));
@@ -94,7 +94,7 @@ impl RTypedData {
     /// # Examples
     ///
     /// ```
-    /// use magnus::{define_class, RTypedData};
+    /// use magnus::{class, define_class, RTypedData};
     /// # let _cleanup = unsafe { magnus::embed::init() };
     ///
     /// #[magnus::wrap(class = "Point")]
@@ -104,7 +104,7 @@ impl RTypedData {
     ///     y: isize,
     /// }
     ///
-    /// let point_class = define_class("Point", Default::default()).unwrap();
+    /// let point_class = define_class("Point", class::object()).unwrap();
     /// let value = RTypedData::wrap(Point { x: 4, y: 2 });
     ///
     /// assert_eq!(value.get::<Point>().unwrap(), &Point { x: 4, y: 2 });