+++ src/typed_data.rs
@@ -320,7 +320,7 @@ where
     /// # Examples
     ///
     /// ```
-    /// use magnus::{prelude::*, define_class, memoize, RClass, TypedData};
+    /// use magnus::{prelude::*, class, define_class, memoize, RClass, TypedData};
     /// # use magnus::DataType;
     ///
     /// struct Example();
@@ -328,7 +328,7 @@ where
     /// unsafe impl TypedData for Example {
     ///     fn class() -> RClass {
     ///         *memoize!(RClass: {
-    ///           let class = define_class("Example", Default::default()).unwrap();
+    ///           let class = define_class("Example", class::object()).unwrap();
     ///           class.undef_alloc_func();
     ///           class
     ///         })
@@ -494,7 +494,7 @@ where
     /// # Examples
     ///
     /// ```
-    /// use magnus::{define_class, prelude::*, typed_data};
+    /// use magnus::{class, define_class, prelude::*, typed_data};
     /// # let _cleanup = unsafe { magnus::embed::init() };
     ///
     /// #[magnus::wrap(class = "Point")]
@@ -503,7 +503,7 @@ where
     ///     y: isize,
     /// }
     ///
-    /// let point_class = define_class("Point", Default::default()).unwrap();
+    /// let point_class = define_class("Point", class::object()).unwrap();
     ///
     /// let value = typed_data::Obj::wrap(Point { x: 4, y: 2 });
     /// assert!(value.is_kind_of(point_class));
@@ -534,7 +534,7 @@ where
     /// # Examples
     ///
     /// ```
-    /// use magnus::{define_class, typed_data};
+    /// use magnus::{class, define_class, typed_data};
     /// # let _cleanup = unsafe { magnus::embed::init() };
     ///
     /// #[magnus::wrap(class = "Point")]
@@ -544,7 +544,7 @@ where
     ///     y: isize,
     /// }
     ///
-    /// let point_class = define_class("Point", Default::default()).unwrap();
+    /// let point_class = define_class("Point", class::object()).unwrap();
     /// let value = typed_data::Obj::wrap(Point { x: 4, y: 2 });
     ///
     /// assert_eq!(&*value, &Point { x: 4, y: 2 });
@@ -634,7 +634,7 @@ where
 /// use std::hash::Hasher;
 ///
 /// use magnus::{
-///     prelude::*, define_class, embed::init, function, gc, method, ruby_handle::RubyHandle, typed_data, value::Opaque,
+///     prelude::*, class, define_class, embed::init, function, gc, method, ruby_handle::RubyHandle, typed_data, value::Opaque,
 ///     DataTypeFunctions, Error, IntoValue, RHash, TypedData, Value,
 /// };
 ///
@@ -697,7 +697,7 @@ where
 ///
 /// let _cleanup = unsafe { init() };
 ///
-/// let class = define_class("Pair", Default::default()).unwrap();
+/// let class = define_class("Pair", class::object()).unwrap();
 /// class
 ///     .define_singleton_method("new", function!(Pair::new, 2))
 ///     .unwrap();
@@ -750,7 +750,7 @@ where
 /// use std::hash::Hasher;
 ///
 /// use magnus::{
-///     prelude::*, define_class, embed::init, function, gc, method, ruby_handle::RubyHandle, typed_data, value::Opaque,
+///     prelude::*, class, define_class, embed::init, function, gc, method, ruby_handle::RubyHandle, typed_data, value::Opaque,
 ///     DataTypeFunctions, Error, IntoValue, RHash, TypedData, Value,
 /// };
 ///
@@ -813,7 +813,7 @@ where
 ///
 /// let _cleanup = unsafe { init() };
 ///
-/// let class = define_class("Pair", Default::default()).unwrap();
+/// let class = define_class("Pair", class::object()).unwrap();
 /// class
 ///     .define_singleton_method("new", function!(Pair::new, 2))
 ///     .unwrap();